split and rename dsl-parser files
